Characters classes restructured
