+++ IlluminatiConfirmed/Game/classes/Game.h
@@ -8,16 +8,15 @@
 #include <vector>
 
 #include "Bullet.h"
-#include "Character.h"
+#include "CharacterSouthPark.h"
 #include "Level.h"
-#include "charactersouthpark.h"
 #include "constants.h"
 
-using IlluminatiConfirmed::Character;
+// using IlluminatiConfirmed::Character;
 
 namespace IlluminatiConfirmed {
 class Game {
- public:
+public:
   Level m_level;
   int m_currentHeroId;
   // b2World *m_world;
@@ -33,7 +32,7 @@ class Game {
 
   void buildBarriers(std::vector<Object> &walls);
 
- public:
+public:
   Game(sf::RenderWindow &window);
   void initNewGame(const std::string &mapFile);
   void initCharacters();
@@ -53,6 +52,8 @@ class Game {
   // void update(time, key);
   void updatePhysics();
 
-  void sendBullet(Character *hero);
+  void sendBullet(BaseCharacter *hero);
+  void saveGame(const std::string &fileName);
+  void loadGame(const std::string &fileName);
 };
 }
