+++ IlluminatiConfirmed/Game/classes/screens/ScreenMenuLoad.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "ScreenMenu.h"
+#include "constants.h"
+
+namespace IlluminatiConfirmed {
+class ScreenMenuLoad : public IlluminatiConfirmed::ScreenMenu {
+public:
+  ScreenMenuLoad();
+  ScreenName run(IlluminatiConfirmed::Game &game, sf::RenderWindow &window);
+};
+}
